% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$..init..}
\alias{AUTO_VI$..init..}
\title{Initialization method}
\arguments{
\item{fitted_mod}{Model. A model object, e.g. \code{lm}.}

\item{keras_mod}{Keras model. A trained computer vision model. Default is
\code{NULL}.}

\item{dat}{Data frame. The data used to fit the model. Default is \code{NULL}.}

\item{node_index}{Integer. An index indicating which node of the output layer
contains the visual signal strength. This is particularly useful
when the keras model has more than one output nodes. Default is \code{1L}.}
}
\value{
Return the object itself.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment. The response variable of this model
is \code{y}.
}
\examples{
my_vi <- auto_vi(fitted_mod = lm(speed ~ dist, data = cars))
my_vi
}
