% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$..str..}
\alias{AUTO_VI$..str..}
\title{String representation of the object}
\value{
A string.
}
\description{
This function returns a string representation of the object.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{AUTO_VI$..str..()
}\if{html}{\out{</div>}}
}
}
\examples{

AUTO_VI$..str..()

my_vi <- auto_vi(fitted_mod = lm(speed ~ dist, data = cars))
my_vi$..str..()
}
