% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$boot_vss}
\alias{AUTO_VI$boot_vss}
\title{Predict visual signal strength for bootstrapped residual plots}
\arguments{
\item{draws}{Integer. Number of simulation draws.}

\item{fitted_mod}{Model. A model object, e.g. \code{lm}.}

\item{keras_mod}{Keras model. A trained computer vision model.}

\item{jitter}{Boolean. Whether to use \code{jitter()} to generate bootstrapped
data instead of sampling from the original data with replacement.}

\item{factor}{Numeric. See also \code{\link[=jitter]{jitter()}}.}

\item{dat}{Data frame. The data used to fit the model.
See also \code{\link[=AUTO_VI$get_dat]{AUTO_VI$get_dat()}}.}

\item{node_index}{Integer. An index indicating which node of the output layer
contains the visual signal strength. This is particularly useful
when the keras model has more than one output nodes.}

\item{keep_boot_dat}{Boolean. Whether to keep the bootstrapped data.}

\item{keep_boot_plot}{Boolean. Whether to keep the bootstrapped plots.}
}
\value{
A tibble with 1 to 3 columns depending on the argument
\code{keep_boot_dat} and \code{keep_boot_plot}.
}
\description{
This function bootstrap the data and refits the model, then
predicts the visual signal strength of the bootstrapped residual plots.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{AUTO_VI$boot_vss(
  draws = 100L,
  fitted_mod = self$fitted_mod,
  keras_mod = self$keras_mod,
  jitter = FALSE,
  factor = 1L,
  dat = self$get_dat(),
  node_index = 1L,
  keep_boot_dat = FALSE,
  keep_boot_plot = FALSE
)
}\if{html}{\out{</div>}}
}
}
