% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$check}
\alias{AUTO_VI$check}
\title{Conduct a auto visual inference check with a computer vision model}
\arguments{
\item{null_draws}{Integer. Number of simulation draws for
\link{AUTO_VI$null_vss}. Default is \code{100L}.}

\item{boot_draws}{Integer. Number of simulation draws for
\link{AUTO_VI$boot_vss}. Default is \code{100L}.}

\item{fitted_mod}{Model. A model object, e.g. \code{lm}.
Default is \code{self$fitted_mod}.}

\item{keras_mod}{Keras model. A trained computer vision model.
Default is \code{self$keras_mod}.}

\item{method}{Function. A method to simulate residuals from the null
hypothesis distribution. For \code{lm}, the recommended method is residual
rotation \link{AUTO_VI$rotate_resid}. Default is \code{self$rotate_resid}.}

\item{dat}{Data frame. The data used to fit the model.
Default is \code{self$get_dat()}.}

\item{node_index}{Integer. An index indicating which node of the output layer
contains the visual signal strength. This is particularly useful
when the keras model has more than one output nodes. Default is \code{1L}.}

\item{keep_dat}{Boolean. Whether to keep the simulated data.
Default is \code{FALSE}.}

\item{keep_plot}{Boolean. Whether to keep the simulated plots.
Default is \code{FALSE}.}
}
\value{
Return the object itself.
}
\description{
This function conducts a visual inference
check with a computer vision model.
}
