% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ_auto_visual_inference.R
\name{AUTO_VI$get_fitted_and_resid}
\alias{AUTO_VI$get_fitted_and_resid}
\title{Get fitted values and residuals out of a model object}
\arguments{
\item{fitted_mod}{Model. A model object, e.g. \code{lm}. Default is
\code{self$fitted_mod}.}
}
\value{
A tibble.
}
\description{
This function gets fitted values and residuals
out of a model object by using \link[stats:fitted.values]{stats::fitted} and \link[stats:residuals]{stats::resid}.
}
\examples{

my_vi <- auto_vi(fitted_mod = lm(speed ~ dist, data = cars))
my_vi$get_fitted_and_resid()
}
