% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$lr_ratio}
\alias{AUTO_VI$lr_ratio}
\title{Compute the likelihood ratio using the simulated result}
\value{
A named vector with three elements \code{boot_likelihood},
\code{null_likelihood} and \code{lr_ratio}.
}
\description{
This function estimates the likelihood
of observing the visual signal strength in terms of the bootstrapped
distribution and the simulated null distribution, and computes the ratio
between these two likelihood.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{AUTO_VI$lr_ratio()
}\if{html}{\out{</div>}}
}
}
