% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$plot_resid}
\alias{AUTO_VI$plot_resid}
\title{Draw a standard residual plot}
\arguments{
\item{data}{Data frame. A data frame containing variables \code{.resid} and
\code{.fitted}. See also \code{\link[=AUTO_VI$get_fitted_and_resid]{AUTO_VI$get_fitted_and_resid()}}.}

\item{theme}{\code{ggtheme}. A \code{ggplot} theme object.
See also \code{\link[ggplot2:ggtheme]{ggplot2::theme_light()}}.}

\item{alpha}{Numeric. Alpha of dot. Value between 0 and 1.}

\item{size}{Numeric. Size of dot. Value between 0 and 1.}

\item{stroke}{Numeric. Stroke of dot. Value between 0 and 1.}

\item{remove_axis}{Boolean. Whether or not to remove the axis.}

\item{remove_legend}{Boolean. Whether or not to remove the legend.}

\item{remove_grid_line}{Boolean. Whether or not to remove the grid lines.}

\item{add_zero_line}{Boolean. Whether or not to add a zero horizontal line.}
}
\value{
A \code{ggplot}.
}
\description{
This function draws a standard residual plot.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{AUTO_VI$plot_resid(
  data = self$get_fitted_and_resid(),
  theme = ggplot2::theme_light(base_size = 11/5),
  alpha = 1,
  size = 0.5,
  stroke = 0.5,
  remove_axis = TRUE,
  remove_legend = TRUE,
  remove_grid_line = TRUE,
  add_zero_line = TRUE
)
}\if{html}{\out{</div>}}
}
}
\examples{

my_vi <- auto_vi(fitted_mod = lm(speed ~ dist, data = cars))
my_vi$plot_resid()
}
\seealso{
\link[visage:VI_MODEL]{visage::VI_MODEL}
}
