% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$remove_plot}
\alias{AUTO_VI$remove_plot}
\title{Remove a plot}
\arguments{
\item{path}{Character. Path to the image file.}

\item{check_ext}{Boolean. Whether to check the file extension.}
}
\value{
Return the object itself.
}
\description{
This function removes a plot.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{AUTO_VI$remove_plot(path, check_ext = TRUE)
}\if{html}{\out{</div>}}
}
}
\examples{

my_vi <- auto_vi(fitted_mod = lm(speed ~ dist, data = cars))
temp_path <- my_vi$save_plot(my_vi$plot_resid())
my_vi$remove_plot(temp_path)
file.exists(temp_path)
}
