% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$save_plot}
\alias{AUTO_VI$save_plot}
\title{Save a plot}
\arguments{
\item{p}{\code{ggplot}. A plot.}

\item{path}{Character. Path to save. See also \code{\link[=tempfile]{tempfile()}}.}

\item{width}{Numeric. Width of the plot.}

\item{height}{Numeric. Height of the plot.}

\item{...}{Arguments passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\value{
A string indicating the path to the file.
}
\description{
This function saves a plot with \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{AUTO_VI$save_plot(
  p,
  path = tempfile(fileext = ".png"),
  width = 7/5,
  height = 7/4,
  ...
)
}\if{html}{\out{</div>}}
}
}
\examples{

my_vi <- auto_vi(fitted_mod = lm(speed ~ dist, data = cars))
my_vi$save_plot(my_vi$plot_resid())
}
