% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$summary_plot}
\alias{AUTO_VI$summary_plot}
\title{Draw a summary plot for the result}
\arguments{
\item{check_result}{List. The result. Default is \code{self$check_result}.}
}
\value{
A \code{ggplot}.
}
\description{
This function draws a summary plot for the result.
}
