% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$unique_obs_correction}
\alias{AUTO_VI$unique_obs_correction}
\title{Compute the adjustment factor for the reduction of the number of
unique observations}
\arguments{
\item{vss}{Numeric. Visual signal strength.}

\item{overlap_ratio.}{Numeric. A value between 0 and 1 indicating the
percentage of original observations presenting in the bootstrapped data.}
}
\value{
A numeric value representing the adjusted visual signal strength.
}
\description{
This function computes the adjustment factor for the reduction
of the number of unique observations. Once we bootstrap the data and refit
the model, the residual plot will usually have less observations due to
sampling with replacement. The visual signal strength of this plot will
also decrease. To correct this effect, we can scale the visual signal
strength based on the unique number of observations in the
bootstrapped data.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{AUTO_VI$unique_obs_correction(vss, overlap_ratio)
}\if{html}{\out{</div>}}
}
}
\examples{
AUTO_VI$unique_obs_correction(1, 2/3)
}
