% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ_auto_visual_inference.R
\docType{data}
\name{AUTO_VI}
\alias{AUTO_VI}
\title{AUTO_VI class environment}
\format{
An environment with S3 class \code{bandicoot_oop}.
}
\usage{
AUTO_VI
}
\description{
This is the class of auto visual inference,
inherited from \link[bandicoot:BASE]{bandicoot::BASE}.
}
\seealso{
Parent class: \link[bandicoot:BASE]{bandicoot::BASE}
\cr
\cr
New attributes: \link{AUTO_VI$check_result}
\cr
\cr
New methods: \link{AUTO_VI$..init..}, \link{AUTO_VI$..str..},
\link{AUTO_VI$get_fitted_and_resid}, \link{AUTO_VI$get_dat},
\link{AUTO_VI$plot_resid}, \link{AUTO_VI$save_plot}, \link{AUTO_VI$remove_plot},
\link{AUTO_VI$vss}, \link{AUTO_VI$rotate_resid}, \link{AUTO_VI$null_vss},
\link{AUTO_VI$boot_vss}, \link{AUTO_VI$check}, \link{AUTO_VI$lr_ratio},
\link{AUTO_VI$summary_plot}
}
\keyword{datasets}
