% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_keras_wrapper.R
\name{KERAS_WRAPPER$..init..}
\alias{KERAS_WRAPPER$..init..}
\title{Initialization method}
\arguments{
\item{keras_mod}{Keras model. A trained computer vision model.}

\item{node_index}{Integer. An index indicating which node of the output layer
contains the visual signal strength. This is particularly useful
when the keras model has more than one output nodes.}
}
\value{
Return the object itself.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{KERAS_WRAPPER$..init..(keras_mod = NULL, node_index = 1L)
}\if{html}{\out{</div>}}
}
}
\examples{
keras_wrapper()

}
