% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_keras_wrapper.R
\name{KERAS_WRAPPER$..str..}
\alias{KERAS_WRAPPER$..str..}
\title{String representation of the object}
\value{
A string.
}
\description{
This function returns a string representation of the object.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{KERAS_WRAPPER$..str..()
}\if{html}{\out{</div>}}
}
}
\examples{

KERAS_WRAPPER$..str..()

wrapper <- keras_wrapper()
wrapper$..str..()
}
