% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_keras_wrapper.R
\name{KERAS_WRAPPER$get_input_height}
\alias{KERAS_WRAPPER$get_input_height}
\title{Get keras model input image height}
\arguments{
\item{keras_model}{Keras model. A trained computer vision model.}
}
\value{
An integer.
}
\description{
This function get the input image height (the input shape
is (batch_size, height, width, channels)) of a keras model.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{KERAS_WRAPPER$get_input_height(keras_model = self$keras_model)
}\if{html}{\out{</div>}}
}
}
\examples{
keras_model <- try(get_keras_model("vss_phn_32"))
if (!inherits(keras_model, "try-error")) {
  keras_wrapper(keras_model)$get_input_height()
}

}
