% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_keras_wrapper.R
\name{KERAS_WRAPPER$get_input_width}
\alias{KERAS_WRAPPER$get_input_width}
\title{Get keras model input image width}
\arguments{
\item{keras_model}{Keras model. A trained computer vision model.}
}
\value{
An integer.
}
\description{
This function get the input image width (the input shape
is (batch_size, height, width, channels)) of a keras model.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{KERAS_WRAPPER$get_input_width(keras_model = self$keras_model)
}\if{html}{\out{</div>}}
}
}
\examples{
if (interactive()) {
  keras_model <- get_keras_model("vss_phn_32")
  keras_model$get_input_width()
}

}
