% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keras_model.R
\name{list_keras_models}
\alias{list_keras_models}
\title{List all available pre-trained computer vision models}
\usage{
list_keras_models()
}
\value{
A tibble of available model names and paths.
}
\description{
This function gets a table of available pre-trained computer vision models
for predicting visual signal strength.
}
\examples{

list_keras_models()

}
