% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{new_class}
\alias{register_method}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bandicoot}{\code{\link[bandicoot]{new_class}}, \code{\link[bandicoot]{register_method}}}
}}

