% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_plot}
\alias{remove_plot}
\title{Remove a plot}
\usage{
remove_plot(path, check_ext = TRUE)
}
\arguments{
\item{path}{Character. Path to the image.}

\item{check_ext}{Boolean. Whether to check the file extension.}
}
\value{
No return. Called for side-effect.
}
\description{
This function removes a plot from a provided path.
}
\examples{
if (interactive()) {
  path <- tempfile(fileext = ".png")
  remove_plot(path)
}

}
