% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_plot}
\alias{save_plot}
\title{Save plot(s)}
\usage{
save_plot(p, path = NULL, width = 7/5, height = 7/4, ...)
}
\arguments{
\item{p}{\code{ggplot}. A plot.}

\item{path}{Character. Path(s) to save the image.}

\item{width}{Numeric. Width of the image.}

\item{height}{Numeric. Height of the image.}

\item{...}{Other arguments passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\value{
The image path(s).
}
\description{
This function save a plot of a list of plots to provided path(s).
}
\examples{
p <- ggplot2::ggplot(cars) + ggplot2::geom_point(ggplot2::aes(dist, speed))
save_plot(p)

}
