% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$..init..}
\alias{AUTO_VI$..init..}
\title{Initialization method}
\arguments{
\item{fitted_model}{Model. A model object, e.g. \code{lm}.}

\item{keras_model}{Keras model. A trained computer vision model.}

\item{data}{Data frame. The data used to fit the model.}

\item{node_index}{Integer. An index indicating which node of the output layer
contains the visual signal strength. This is particularly useful
when the keras model has more than one output nodes.}
}
\value{
Return the object itself.
}
\description{
This function will be called after an instance is built. User
input will be stored in the environment.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{AUTO_VI$..init..(fitted_model, keras_model = NULL, data = NULL, node_index = 1L)
}\if{html}{\out{</div>}}
}
}
\examples{
my_vi <- auto_vi(fitted_model = lm(speed ~ dist, data = cars))
my_vi
}
