% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$boot_vss}
\alias{AUTO_VI$boot_vss}
\title{Predict visual signal strength for bootstrapped residual plots}
\arguments{
\item{draws}{Integer. Number of simulation draws. Default is \code{100L}.}

\item{fitted_mod}{Model. A model object, e.g. \code{lm}.
Default is \code{self$fitted_mod}.}

\item{keras_mod}{Keras model. A trained computer vision model.
Default is \code{self$keras_mod}.}

\item{dat}{Data frame. The data used to fit the model.
Default is \code{self$get_dat()}.}

\item{node_index}{Integer. An index indicating which node of the output layer
contains the visual signal strength. This is particularly useful
when the keras model has more than one output nodes. Default is \code{1L}.}

\item{keep_boot_dat}{Boolean. Whether to keep the bootstrapped data.
Default is \code{FALSE}.}

\item{keep_boot_plot}{Boolean. Whether to keep the bootstrapped plots.
Default is \code{FALSE}.}
}
\value{
A tibble with 1 to 3 columns depending on the argument
\code{keep_boot_dat} and \code{keep_boot_plot}.
}
\description{
This function bootstrap the data and refits the model, then
predicts the visual signal strength of the bootstrapped residual plots.
}
