% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$check}
\alias{AUTO_VI$check}
\title{Conduct a auto visual inference check with a computer vision model}
\arguments{
\item{null_draws}{Integer. Number of simulation draws for
\code{\link[=AUTO_VI$null_vss]{AUTO_VI$null_vss()}}.}

\item{boot_draws}{Integer. Number of simulation draws for
\code{\link[=AUTO_VI$boot_vss]{AUTO_VI$boot_vss()}}.}

\item{fitted_mod}{Model. A model object, e.g. \code{lm}.}

\item{keras_mod}{Keras model. A trained computer vision model.}

\item{null_method}{Function. A method to simulate residuals from the null
hypothesis distribution. For \code{lm}, the recommended method is residual
rotation \code{\link[=AUTO_VI$rotate_resid]{AUTO_VI$rotate_resid()}}.}

\item{jitter}{Boolean. Whether to use \code{jitter()} to generate bootstrapped
data instead of sampling from the original data with replacement.}

\item{factor}{Numeric. See also \code{\link[=jitter]{jitter()}}.}

\item{dat}{Data frame. The data used to fit the model.
See also \code{\link[=AUTO_VI$get_dat]{AUTO_VI$get_dat()}}.}

\item{node_index}{Integer. An index indicating which node of the output layer
contains the visual signal strength. This is particularly useful
when the keras model has more than one output nodes.}

\item{keep_dat}{Boolean. Whether to keep the simulated data.}

\item{keep_plot}{Boolean. Whether to keep the simulated plots.}
}
\value{
Return the object itself.
}
\description{
This function conducts a visual inference
check with a computer vision model. The result will be stored in
\code{self$check_result}.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{AUTO_VI$check(
  null_draws = 100L,
  boot_draws = 100L,
  fitted_mod = self$fitted_mod,
  keras_mod = self$keras_mod,
  null_method = self$rotate_resid,
  jitter = FALSE,
  factor = 1L,
  dat = self$get_dat(),
  node_index = self$node_index,
  keep_dat = FALSE,
  keep_plot = FALSE
)
}\if{html}{\out{</div>}}
}
}
