% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$get_dat}
\alias{AUTO_VI$get_dat}
\title{Get data out of a model object}
\arguments{
\item{fitted_mod}{Model. A model object, e.g. \code{lm}. Default is
\code{self$fitted_mod}.}
}
\value{
A tibble.
}
\description{
This function gets the data out of a model object
by using \link[stats:model.frame]{stats::model.frame} if \code{self$dat} is \code{NULL}.
}
\examples{

my_vi <- auto_vi(fitted_mod = lm(speed ~ dist, data = cars))
my_vi$get_dat()
}
