% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZZ_auto_visual_inference.R
\name{AUTO_VI$null_vss}
\alias{AUTO_VI$null_vss}
\title{Simulate null plots and predict the visual signal strength}
\arguments{
\item{draws}{Integer. Number of simulation draws. Default is \code{100L}.}

\item{fitted_mod}{Model. A model object, e.g. \code{lm}.
Default is \code{self$fitted_mod}.}

\item{keras_mod}{Keras model. A trained computer vision model.
Default is \code{self$keras_mod}.}

\item{method}{Function. A method to simulate residuals from the null
hypothesis distribution. For \code{lm}, the recommended method is residual
rotation \link{AUTO_VI$rotate_resid}. Default is \code{self$rotate_resid}.}

\item{node_index}{Integer. An index indicating which node of the output layer
contains the visual signal strength. This is particularly useful
when the keras model has more than one output nodes. Default is \code{1L}.}

\item{keep_null_dat}{Boolean. Whether to keep the simulated null data.
Default is \code{FALSE}.}

\item{keep_null_plot}{Boolean. Whether to keep the simulated null plots.
Default is \code{FALSE}.}
}
\value{
A tibble with 1 to 3 columns depending on the argument
\code{keep_null_dat} and \code{keep_null_plot}.
}
\description{
This function simulates null plots from the null hypothesis
distribution, and predicts the visual signal strength.
}
