% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$p_value}
\alias{AUTO_VI$p_value}
\title{Compute the p-value based on the check result}
\arguments{
\item{type}{Character. Either "null" or "boot".}
}
\value{
A numeric value representing the desired p-value.
}
\description{
This function computes the p-value of observing the
visual signal strength of the original residual plot based on the null
distribution, and computes the p-value of observing the mean of the
bootstrapped distribution based on the null distribution.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{AUTO_VI$p_value(type = "null")
}\if{html}{\out{</div>}}
}
}
