% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$plot_resid}
\alias{AUTO_VI$plot_resid}
\title{Draw a standard residual plot}
\arguments{
\item{x}{Data frame. A data frame containing variables \code{.resid} and
\code{.fitted}. Default is \code{self$get_fitted_and_resid()}.}

\item{theme}{\code{ggtheme}. A \code{ggplot} theme object.
Default is \code{ggplot2::theme_light(base_size = 11/5)}.}

\item{alpha}{Numeric. Alpha of dot. Value between 0 and 1. Default is \code{1}.}

\item{size}{Numeric. Size of dot. Value between 0 and 1. Default is \code{0.5}.}

\item{stroke}{Numeric. Stroke of dot. Value between 0 and 1.
Default is \code{0.5}.}

\item{remove_axis}{Boolean. Whether or not to remove the axis.
Default is \code{TRUE}.}

\item{remove_legend}{Boolean. Whether or not to remove the legend.
Default is \code{TRUE}.}

\item{remove_grid_line}{Boolean. Whether or not to remove the grid lines.
Default is \code{TRUE}.}

\item{add_zero_line}{Boolean. Whether or not to add a zero horizontal line.
Default is \code{TRUE}.}
}
\value{
A \code{ggplot}.
}
\description{
This function draws a standard residual plot.
}
\examples{

my_vi <- auto_vi(fitted_mod = lm(speed ~ dist, data = cars))
my_vi$plot_resid()
}
\seealso{
\link[visage:VI_MODEL]{visage::VI_MODEL}
}
