% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$save_plot}
\alias{AUTO_VI$save_plot}
\title{Save a plot}
\arguments{
\item{p}{\code{ggplot}. A plot.}

\item{path}{Character. Path to save. Default is \code{tempfile(fileext = '.png')}.}

\item{width}{Numeric. Width of the plot. Default is \code{7/5}.}

\item{height}{Numeric. Height of the plot. Default is \code{7/4}.}

\item{...}{Arguments passed to \link[ggplot2:ggsave]{ggplot2::ggsave}.}
}
\value{
A string indicating the path to the file.
}
\description{
This function saves a plot with \link[ggplot2:ggsave]{ggplot2::ggsave}.
}
\examples{

my_vi <- auto_vi(fitted_mod = lm(speed ~ dist, data = cars))
my_vi$save_plot(my_vi$plot_resid())
}
