% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$summary_plot}
\alias{AUTO_VI$summary_plot}
\title{Draw a summary plot for the result}
\value{
A \code{ggplot}.
}
\description{
This function draws a summary plot for the result.
\subsection{Usage}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{AUTO_VI$summary_plot()
}\if{html}{\out{</div>}}
}
}
