% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_auto_visual_inference.R
\name{AUTO_VI$vss}
\alias{AUTO_VI$vss}
\title{Predict the visual signal strength}
\arguments{
\item{p}{\code{ggplot}/List. A \code{ggplot} or a list of \code{ggplot}.}

\item{keras_mod}{Keras model. A trained computer vision model.
Default is \code{self$keras_mod}.}

\item{node_index}{Integer. An index indicating which node of the output layer
contains the visual signal strength. This is particularly useful
when the keras model has more than one output nodes. Default is \code{1L}.}
}
\value{
A numeric vector.
}
\description{
This function predicts the visual signal strength.
}
