% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_plot}
\alias{remove_plot}
\title{Remove a plot}
\usage{
remove_plot(path, check_ext = TRUE)
}
\arguments{
\item{path}{Character. Path to the image.}

\item{check_ext}{Boolean. Whether to check the file extension.}
}
\value{
No return. Called for side-effect.
}
\description{
This function removes a plot from a provided path.
}
\examples{
p <- ggplot2::ggplot(cars) + ggplot2::geom_point(ggplot2::aes(dist, speed))
path <- save_plot(p)
remove_plot(path)

}
