% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot}
\usage{
save_plot(p, path = NULL, width = 7/5, height = 7/4, ...)
}
\arguments{
\item{p}{GGplot. A plot.}

\item{path}{Character. Path to save the image.}

\item{width}{Numeric. Width of the image.}

\item{height}{Numeric. Height of the image.}

\item{...}{Other arguments passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\value{
The image path.
}
\description{
This function save a plot to a provided path.
}
\examples{
if (interactive()) {
  p <- ggplot2::ggplot(cars) + ggplot2::geom_point(ggplot2::aes(dist, speed))
  save_plot(p)
}

}
